@@ -1,7 +1,5 @@
 import range from 'lodash/range';
 
-import { currentVersion } from '../config/global';
-import type { LiturgyDocument } from '../types';
 import version2 from './migrations/version2';
 import version3 from './migrations/version3';
 import version4 from './migrations/version4';
@@ -9,6 +7,8 @@ import version5 from './migrations/version5';
 import version6 from './migrations/version6';
 import version7 from './migrations/version7';
 import version8 from './migrations/version8';
+import { currentVersion } from '../config/global';
+import type { LiturgyDocument } from '../types';
 
 function toVersion(version: number, data: any) {
   switch (version) {
