@@ -1,4 +1,4 @@
-/// <reference types="react-scripts" />
+/// <reference types="vite/client" />
 
 declare module 'superagent-jsonp';
 
@@ -11,4 +11,12 @@ declare module '@mui/material/styles' {
       border: string;
     };
   }
+}
+
+interface ImportMetaEnv {
+  readonly VITE_SLIDESHOW_ENABLED: string;
+}
+
+interface ImportMeta {
+  readonly env: ImportMetaEnv;
 }
